qcacld-3.0: Update correct RSSI for connected case getstainfo
When a station is connected to SAP and host driver receives
getstainfo command, it gives rssi info as part of tihs command.
Currently host driver does not update this rssi information
correctly because of which userspace is getting rssi info as
positive value which is not correct.

To address above issue, update rssi info correctly in the host
driver.

Change-Id: I87033184644a1d246f8669c74153c46837a35276
CRs-Fixed: 3184344
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>