qcacld-3.0: Add new ini to allow STA-SAP scc on indoor only channel
Add new ini for STA-SAP scc on 5 GHz indoor only channel -
"sta_sap_scc_on_indoor_chan"

When this ini is enabled, SAP moves to the STA channel
when STA is connected on indoor only 5 GHz channel with the
AP. When this ini is disabled, SAP moves to any 2.4 GHz
channel.
When gindoor_channel_support=1, this ini will not be considered and
SAP can come up on indoor channel.

Change-Id: Ia3ebc6d7fc2e6e569cde8e8a8b38ca76036b8fda
CRs-Fixed: 3186938
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>