qcacmn: Mark 6 GHz indoor only channels as NO_IR
As part of the sta_sap_scc_on_indoor support, a concurrent SAP
without indoor channel support can be started on an indoor
channel, therefore NO_IR flag is not set for the indoor
channels in the master channel list.

However, in countries where VLP is not supported, the concurrent
SAP is not allowed on these non-VLP channels. Therefore, restrict
the SCC indoor support only for INDOOR_ONLY 5 GHz frequencies.

Change-Id: I29f7bc4bed5fe5d06d535126eb8958425b5ad313
CRs-Fixed: 3208199
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>