qcacld-3.0: Modify hardware name to static char array
Currently hardware name pointer is present in hdd context. It's
allocated memory during wlan start modules and hence destroyed
during the stop modules. This info is required when wifi is
turned off.

In order to address this issue, declare hardware name as a static
char array in hdd context structure. And modify it's usage
accordingly.

Change-Id: I4663acd602947a704daccfee4da78c207be095f8
CRs-Fixed: 3288897
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>