msm: ipa3: add new mutex lock to update is_modem_up flag.
Adding new mutex lock to update is_modem_up flag because
in Modem SSR scenarios if we use mutex variable "lock" it
is waiting for it to be released by other thread & partially
leading to SSR timeout. As we don't need to have a common
mutex variable "lock" for this, a new mutex can help in moving
further in clean up the pipes.

Change-Id: I8e6e4c0181f7765b334aefefd9616329503fdc2c
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>