qcacld-3.0: Allow suspend in Deep Sleep/Hibernate in wearables
Deep Sleep or Hibernate are system level mode power
modes wearable targets in which wlan firmware is shutdown
gracefully and linux kernel and platform are suspended.

Currently, wlan suspend and pm_suspend is not allowed if
firmware is down, with this change suspend is allowed if firmware
is down due to Deep Sleep or Hibernate low power state.

Change-Id: Ide7b8957d1cfe455d57769c30f0bb3d6e564cddc
CRs-Fixed: 3108081
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>