qcacld-3.0: Use dynamic symbolic links for multi-ko support
This reverts Change-Id I7d75bb6a82fd27d8925d53529b4252b782457bc8
("qcacld-3.0: Use static symbolic links for multi-ko support") along
with removing per-chipset symbolic links added after that patch.

The original patch was created to allow the qcacld-3.0 driver to
conform to the Hermetic Build feature which required that the source
directory not be modified during the build process. However,
qcacld-3.0 is transitioning to the Bazel DDK, and the presence of the
symbolic links is interfering with Bazel. In addition, the Bazel
build system won't use these symbolic links. So the support both the
legacy build system (in a non-hermetic manner) as well as the Bazel
DDK (in a hermetic manner), revert back to dynamically creating the
symbolic links.

Change-Id: I2b6a0ed19f7f4b3ae31d28e2b4683afd0671f19a
CRs-Fixed: 3495618
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>