qcacld-3.0: Enable netdev queues before sending ROAM_SYNC_COMPLETE
Upon the completion of roam sync indication event, the
roam sync complete cmd is first sent to the firmware.

This is followed by the netdev queues enable. This
approach adds latency in the first downstream tx frame,
and the FT roaming KPI is not met due to this latency.

To fix this, if roam sync handling is successful, enable
the queues and then send roam sync complete to the firmware.

It's observed that the current change improves roaming
latency by 6 msec.

Change-Id: Ib05ca01892b6d8b62badcafb64bb8c16270ee9f8
CRs-Fixed: 3483751
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>