qcacld-3.0: Use proper vdev ID while processing the OP mode update
While enqueuing the WMA_UPDATE_OP_MODE, session ID is filled with
the index where the BSSID entry matched in PE session. And after
dequeuing it this session ID is used as vdev ID and sent to FW in
WMI_PEER_SET_PARAM_CMDID, which results in sending an update for
an invalid vdev ID.
Hence use the smeSessionId value from PE session.

Change-Id: I1b550be2c98c3eb0f3716a953408ee2c217e8d9b
CRs-Fixed: 3510396
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>