qcacld-3.0: Fix an array size issue
Currently, a MACRO is defined used to record the array size. However,
This has a potential risk if the array size changes but the corresponding
ARRAY_SIZE MACRO is not modified accordingly. To solve this issue, this
change uses QDF_ARRAY_SIZE(_arr) to express the array size, so that the
value of QDF_ARRAY_SIZE(_arr) will always represent the size of the array.

Change-Id: I7108a21a91224fe25cfac95dc5801f83aee0a651
CRs-Fixed: 3501175
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>