qcacld-3.0: Refactor fixed macros from Kbuild
In some build environments the length of the C compiler command line
exceeds the maximum length of the underlying shell, resulting in the
shell reporting: Argument list too long.

Kbuild has many definitions which add -D switches to the C compiler
command line to define macros to be used for conditional compilation.
Most of these are set conditionally, but there are some that are set
unconditionally.

As one step in the process of shortening the command line, refactor
the macros that are set unconditionally out of Kbuild, and put them in
a separate .h file which is then included in every build. As long as
the added -include switch takes up less room than the sum of the -D
switches, then this will result in a shorter command line.

Change-Id: I95e0e8237d15dee58e519ea526a74c52eeb7dbd2
CRs-Fixed: 3460064
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>