qcacld-3.0: Configure join timeout based on beacon interval
Join timeout can be configured to 10 times the beacon interval
of the candidate beacon if it is in bounds with the join failure
timeout ini. However, host always configures join failure timer
with the default join failure timeout.

Fix the join failure timeout calculation by considering the
beacon interval.

CRs-Fixed: 3628892
Change-Id: I6c238c7db8ad6925c11fa97b3a698643187fec95
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>