qcacmn: Fix NULL pointer dereference while fetching WEP key
During WEP connection if the user selects key index other
than 1 on AP. Then in STA the WEP key is stored at that selected
key index and after the association during key install the WEP
key is fetched from a hardcoded key index '1', which results in
fetching a NULL pointer. But there is no proper NULL check to
handle such cases, which leads to a crash.
Add a NULL check before using that pointer for fetching WEP key.

Change-Id: I1111ac76b5d692f651ebeff573a7d9f1bb83b66f
CRs-Fixed: 3860734
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>