qcacmn: Use SP channels for passive scan
Currently host driver doesn't consider SP channels for scan.
So if any country supports only SP power type then STA is not
able to scan and connect to the AP which is operating on
SP channel.

To address this issue enable passive scanning for SP only
channels and add logic to prevent turning up SAP on these
channels.

Change-Id: Ieb14ded41c67e23e066c8a0a5c5e450975d8c4c2
CRs-Fixed: 3529200
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>