qcacld-3.0: Disable DDK bzl compilation for monaco older kernels
Disable bzl compilation for older 4.19 and 5.15 kernels for monaco
target as it is using Kbuild on LW.UM.2.3 SI for compilation.
The issue is LW.UM.2.3 SI is using older kernel but as bzl is enabled
by default, getting compilation errors when reading the bzl file's
soc-repo flag. So as a result, disable bzl compilation for older
kernels so that Kbuild is used.

Change-Id: Ie2208f4f18ab82b6aa08773e04446c0fc89a4f36
CRs-Fixed: 4187906
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>